Update size and dimensions of landing image
Previously, when the user reduced screen size below 1000 pixels, the landing
image would not take the full height of the background.  This resulted in the
color overlay overflowing, which looked like a band of blue underneath the
image.

Fix this issue by updating the dimensions of the landing image so it has a
larger height.

This has also resulted in the size of the image being smaller, which should
help with loading time.